fix(control-plane): cast reaper sweep operands to int (Postgres 42725)
claimDevicesToNotifyLost summed two untyped bound params as ($1 + $2)::int, which Postgres rejects with 42725 "operator is not unique: unknown + unknown" — the ::int cast applies to the result, not the operands. The reaper sweep has thrown every ~10s since deb88f0, silently disabling the device "lost connection" notice (stale-session reaping still ran). Cast each operand: ($1::int + $2::int). Verified against the live DB: broken form errors, cast form returns the cutoff.

+++ apps/control-plane/src/db/repo.ts
@@ -679,7 +679,7 @@ export async function claimDevicesToNotifyLost(
           SELECT 1 FROM sessions s
            WHERE s.device_id = d.id
              AND (s.state <> 'ended'
-                  OR s.last_event_at >= now() - (($1 + $2)::int * interval '1 second'))
+                  OR s.last_event_at >= now() - (($1::int + $2::int) * interval '1 second'))
         )
       RETURNING d.id, d.account_id, d.hostname, d.os, (d.afk = 'on') AS notify`,
     [graceSeconds, staleSeconds],
